+++ fvm/tests/fvm.rs
@@ -1,4 +1,4 @@
 mod default_kernel;
 mod dummy;
 
-use dummy::DummyExterns;
+use dummy::*;
